+++ src/test/java/com/jomm/terroir/business/model/TestCategory.java
@@ -42,7 +42,7 @@ public class TestCategory {
 	@Test
 	public final void testGetterSetter() {
 		String test = "test";
-		Long nb = (long) 0;
+		Long nb = 0L;
 		
 		// Id
 		category.setId(nb);
