+++ src/test/java/com/jomm/terroir/business/model/TestSite.java
@@ -43,7 +43,7 @@ public class TestSite {
 	@Test
 	public final void testGetterSetter() {
 		String test = "test";
-		Long nb = (long) 0;
+		Long nb = 0L;
 		
 		// Id
 		site.setId(nb);
